refactor: api 구조 수정

+++ RunMate/data/src/main/java/com/D107/runmate/data/remote/response/history/MyRunItem.kt
@@ -8,13 +8,13 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 data class MyRunItem(
     val addedToCourse: Boolean,
-    val avgBpm: Int,
-    val avgPace: Int,
-    val calories: Int,
+    val avgBpm: Double,
+    val avgPace: Double,
+    val calories: Double,
     val courseLiked: Boolean,
     val courseLikes: Int,
-    val distance: Int,
-    val time: Int
+    val distance: Double,
+    val time: Long
 ): BaseResponse {
     companion object: DataMapper<MyRunItem, MyRun> {
         override fun MyRunItem.toDomainModel(): MyRun {
