+++ RunMate/domain/src/main/java/com/D107/runmate/domain/model/history/MyRun.kt
@@ -4,11 +4,11 @@ import com.D107.runmate.domain.model.base.BaseModel
 
 data class MyRun(
     val addedToCourse: Boolean,
-    val avgBpm: Int,
-    val avgPace: Int,
-    val calories: Int,
+    val avgBpm: Double,
+    val avgPace: Double,
+    val calories: Double,
     val courseLiked: Boolean,
     val courseLikes: Int,
-    val distance: Int,
-    val time: Int
+    val distance: Double,
+    val time: Long
 ): BaseModel
